CalendarScreen.kt:  - Removed inline calendar grid logic (weekday row, day calculations, and day rendering)  - Removed inline schedule section UI  - Removed inline classes section UI  - Removed inline assignment dialog implementation  - Removed inline class dialog implementation  - Removed duplicated reset logic blocks and replaced with helper functions  - Cleaned unused imports after refactor  - Refactored into a coordinator screen that delegates UI to separate composables
AssignmentDialog.kt:

 - Created new composable for assignment creation and editing dialog
 - Added class selection dropdown using exposed dropdown menu
 - Added assignment name input field
 - Added due time display and time picker trigger
 - Added support for create, update, and delete actions
 - Added validation handling through enabled confirm button

CalendarGrid.kt:

 - Created new composable for rendering monthly calendar grid
 - Added weekday header row (Su–Sa)
 - Implemented day cell layout with consistent alignment
 - Added current day highlighting
 - Added selected date highlighting
 - Added indicator dots for dates with assignments
 - Implemented date click handling (select vs reopen dialog)

CalendarHeader.kt:

 - Created new composable for calendar header
 - Added month/year display formatting
 - Added navigation controls for previous and next month
 - Added back button handling

ClassDialog.kt:

 - Created new composable for class creation and editing dialog
 - Added class name input field
 - Added meeting day selection with checkboxes
 - Added start and end time inputs with time picker triggers
 - Added support for create, update, and delete actions
 - Enabled scrolling for dialog content to handle smaller screens

ClassesSection.kt:

 - Created new composable for displaying classes for selected date
 - Added conditional rendering for empty state and selected date requirement
 - Added class cards with name, time range, and meeting days
 - Added click handling to trigger class editing

ScheduleSection.kt:

 - Created new composable for displaying assignments for selected date
 - Added conditional rendering for empty state and selected date requirement
 - Added schedule item cards with class name, assignment name, and due time
 - Added click handling to trigger assignment editing

TimePickerUtils.kt:

 - Created utility file for time picker logic
 - Added function to format time into 12-hour AM/PM format
 - Added reusable function to launch Android TimePickerDialog
 - Removed time picker logic from CalendarScreen.kt and centralized it for reuse

 I refactored it :)
