+++ app/src/main/java/com/example/cinet/ClassDialog.kt
@@ -1,0 +1,133 @@
+package com.example.cinet
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Checkbox
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun ClassDialog(
+    editingClass: ClassItem?,
+    className: String,
+    onClassNameChange: (String) -> Unit,
+    classStartTime: String,
+    classEndTime: String,
+    selectedMeetingDays: Set<String>,
+    onMeetingDaysChange: (Set<String>) -> Unit,
+    weekdayOptions: List<String>,
+    onPickStartTime: () -> Unit,
+    onPickEndTime: () -> Unit,
+    onDismiss: () -> Unit,
+    onConfirm: () -> Unit,
+    onDelete: (() -> Unit)?
+) {
+    AlertDialog(
+        onDismissRequest = onDismiss,
+        title = {
+            Text(if (editingClass == null) "Create Class" else "Edit Class")
+        },
+        text = {
+            Column(
+                modifier = Modifier.verticalScroll(rememberScrollState())
+            ) {
+                OutlinedTextField(
+                    value = className,
+                    onValueChange = onClassNameChange,
+                    label = { Text("Class Name") },
+                    modifier = Modifier.fillMaxWidth()
+                )
+
+                Spacer(modifier = Modifier.height(12.dp))
+                Text("Meeting Days")
+
+                weekdayOptions.forEach { day ->
+                    Row(
+                        modifier = Modifier.fillMaxWidth(),
+                        verticalAlignment = androidx.compose.ui.Alignment.CenterVertically
+                    ) {
+                        Checkbox(
+                            checked = selectedMeetingDays.contains(day),
+                            onCheckedChange = { checked ->
+                                onMeetingDaysChange(
+                                    if (checked) selectedMeetingDays + day
+                                    else selectedMeetingDays - day
+                                )
+                            }
+                        )
+                        Text(day)
+                    }
+                }
+
+                Spacer(modifier = Modifier.height(8.dp))
+
+                OutlinedTextField(
+                    value = classStartTime,
+                    onValueChange = {},
+                    label = { Text("Start Time") },
+                    modifier = Modifier.fillMaxWidth(),
+                    readOnly = true
+                )
+
+                Spacer(modifier = Modifier.height(8.dp))
+
+                Button(onClick = onPickStartTime) {
+                    Text("Pick Start Time")
+                }
+
+                Spacer(modifier = Modifier.height(12.dp))
+
+                OutlinedTextField(
+                    value = classEndTime,
+                    onValueChange = {},
+                    label = { Text("End Time") },
+                    modifier = Modifier.fillMaxWidth(),
+                    readOnly = true
+                )
+
+                Spacer(modifier = Modifier.height(8.dp))
+
+                Button(onClick = onPickEndTime) {
+                    Text("Pick End Time")
+                }
+            }
+        },
+        confirmButton = {
+            Button(onClick = onConfirm) {
+                Text(if (editingClass == null) "Save Class" else "Update Class")
+            }
+        },
+        dismissButton = {
+            Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
+                if (onDelete != null) {
+                    Button(
+                        onClick = onDelete,
+                        colors = ButtonDefaults.buttonColors(
+                            containerColor = MaterialTheme.colorScheme.error
+                        )
+                    ) {
+                        Text("Delete")
+                    }
+                }
+
+                OutlinedButton(onClick = onDismiss) {
+                    Text("Cancel")
+                }
+            }
+        }
+    )
+}
